requirements.txt
flask
pandas
numpy
openai
matplotlib
python-dotenv
gunicorn